refactor(frontend): Rename component files to PascalCase and reorganize structure
