+++ app/settings/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Settings } from "@/components/settings";
+import { Settings } from "@/components/Settings";
 
 export default function SettingsPage() {
   return <Settings />;
