+++ components/CreateDatabaseDialog.tsx
@@ -17,7 +17,7 @@ import { createDatabase } from "@/lib/tauri";
 import { useToast } from "@/components/ui/use-toast";
 import { open } from "@tauri-apps/plugin-dialog";
 import { saveLastDatabasePath } from "@/lib/storage";
-import { PasswordStrengthMeter } from "@/components/password-strength-meter";
+import { PasswordStrengthMeter } from "@/components/PasswordStrengthMeter";
 
 interface CreateDatabaseDialogProps {
   isOpen: boolean;
