+++ components/KdfWarningDialog.tsx
