+++ components/PasswordStrengthMeter.tsx
