+++ components/UnlockScreen.tsx
@@ -8,8 +8,8 @@ import { FolderOpen, Plus } from "lucide-react";
 import { openDatabase } from "@/lib/tauri";
 import { useToast } from "@/components/ui/use-toast";
 import { open } from "@tauri-apps/plugin-dialog";
-import { CreateDatabaseDialog } from "@/components/create-database-dialog";
-import { KdfWarningDialog } from "@/components/kdf-warning-dialog";
+import { CreateDatabaseDialog } from "@/components/CreateDatabaseDialog";
+import { KdfWarningDialog } from "@/components/KdfWarningDialog";
 import { saveLastDatabasePath } from "@/lib/storage";
 import { invoke } from "@tauri-apps/api/core";
 import Image from "next/image";
