+++ components/group-tree/index.tsx
@@ -1,0 +1,323 @@
+"use client";
+
+import { useState, useEffect } from "react";
+import { Button } from "@/components/ui/button";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import { Plus, Star, LayoutDashboard } from "lucide-react";
+import { createGroup, deleteGroup, renameGroup, moveGroup } from "@/lib/tauri";
+import { useToast } from "@/components/ui/use-toast";
+import type { GroupData } from "@/lib/tauri";
+import { ask } from "@tauri-apps/plugin-dialog";
+import {
+  ContextMenu,
+  ContextMenuContent,
+  ContextMenuItem,
+  ContextMenuTrigger,
+} from "@/components/ui/context-menu";
+import { getIconComponent } from "@/components/IconPicker";
+import { saveGroupTreeState } from "@/lib/group-state";
+import {
+  DndContext,
+  DragOverlay,
+  useSensor,
+  useSensors,
+  PointerSensor,
+  DragStartEvent,
+  DragOverEvent,
+  DragEndEvent,
+  closestCenter,
+} from "@dnd-kit/core";
+
+import { DraggableFolder } from "./DraggableFolder";
+import { CreateGroupDialog, RenameGroupDialog } from "./GroupDialogs";
+import { findGroupByUuid, isDescendant } from "./utils";
+import type { GroupTreeProps } from "./types";
+
+export function GroupTree({
+  group,
+  selectedUuid,
+  onSelectGroup,
+  onRefresh,
+  onGroupDeleted,
+  dbPath,
+  initialExpandedGroups,
+}: GroupTreeProps) {
+  const [expandedGroups, setExpandedGroups] = useState<Set<string>>(
+    initialExpandedGroups || new Set([group.uuid])
+  );
+  const [showCreateDialog, setShowCreateDialog] = useState(false);
+  const [showRenameDialog, setShowRenameDialog] = useState(false);
+  const [newGroupName, setNewGroupName] = useState("");
+  const [newGroupIconId, setNewGroupIconId] = useState(48);
+  const [renameGroupUuid, setRenameGroupUuid] = useState("");
+  const [renameGroupName, setRenameGroupName] = useState("");
+  const [renameGroupIconId, setRenameGroupIconId] = useState(48);
+  const [parentUuid, setParentUuid] = useState<string | null>(null);
+  const [activeId, setActiveId] = useState<string | null>(null);
+  const [overId, setOverId] = useState<string | null>(null);
+  const { toast } = useToast();
+
+  useEffect(() => {
+    saveGroupTreeState(dbPath, expandedGroups, selectedUuid);
+  }, [expandedGroups, selectedUuid, dbPath]);
+
+  const sensors = useSensors(
+    useSensor(PointerSensor, {
+      activationConstraint: { distance: 8 },
+    })
+  );
+
+  useEffect(() => {
+    if (activeId) {
+      document.body.style.cursor = 'grabbing';
+    } else {
+      document.body.style.cursor = '';
+    }
+    return () => { document.body.style.cursor = ''; };
+  }, [activeId]);
+
+  const toggleExpand = (uuid: string) => {
+    const newExpanded = new Set(expandedGroups);
+    if (newExpanded.has(uuid)) {
+      newExpanded.delete(uuid);
+    } else {
+      newExpanded.add(uuid);
+    }
+    setExpandedGroups(newExpanded);
+  };
+
+  const handleCreateGroup = async () => {
+    if (!newGroupName.trim()) return;
+
+    try {
+      await createGroup(newGroupName, parentUuid, newGroupIconId);
+      toast({ title: "Success", description: "Group created successfully", variant: "success" });
+      setNewGroupName("");
+      setNewGroupIconId(48);
+      setShowCreateDialog(false);
+      onRefresh();
+    } catch (error: any) {
+      toast({
+        title: "Error",
+        description: typeof error === 'string' ? error : (error?.message || "Failed to create group"),
+        variant: "destructive",
+      });
+    }
+  };
+
+  const handleRenameGroup = async () => {
+    if (!renameGroupName.trim()) return;
+
+    try {
+      await renameGroup(renameGroupUuid, renameGroupName, renameGroupIconId);
+      toast({ title: "Success", description: "Group renamed successfully", variant: "success" });
+      setShowRenameDialog(false);
+      setRenameGroupName("");
+      setRenameGroupUuid("");
+      onRefresh();
+    } catch (error: any) {
+      toast({
+        title: "Error",
+        description: typeof error === 'string' ? error : (error?.message || "Failed to rename group"),
+        variant: "destructive",
+      });
+    }
+  };
+
+  const handleDeleteGroup = async (uuid: string) => {
+    const groupToDelete = findGroupByUuid(group, uuid);
+    const groupName = groupToDelete?.name || "this group";
+    
+    const shouldDelete = await ask(
+      `Are you sure you want to delete "${groupName}" and all its contents?`,
+      { kind: "warning", title: "Delete Group" }
+    );
+    
+    if (!shouldDelete) return;
+
+    try {
+      await deleteGroup(uuid);
+      toast({ title: "Success", description: "Group deleted successfully", variant: "success" });
+      if (onGroupDeleted) onGroupDeleted(uuid);
+      onRefresh();
+    } catch (error: any) {
+      toast({
+        title: "Error",
+        description: typeof error === 'string' ? error : (error?.message || "Failed to delete group"),
+        variant: "destructive",
+      });
+    }
+  };
+
+  const handleDragStart = (event: DragStartEvent) => {
+    setActiveId(event.active.id as string);
+  };
+
+  const handleDragOver = (event: DragOverEvent) => {
+    const { active, over } = event;
+    if (!over || active.id === over.id) {
+      if (overId !== null) setOverId(null);
+      return;
+    }
+    if (overId !== over.id) setOverId(over.id as string);
+  };
+
+  const handleDragEnd = async (event: DragEndEvent) => {
+    const { active, over } = event;
+    setActiveId(null);
+    setOverId(null);
+
+    if (!over || active.id === over.id) return;
+
+    const draggedId = active.id as string;
+    const targetId = over.id as string;
+    const draggedGroup = findGroupByUuid(group, draggedId);
+    const targetGroup = findGroupByUuid(group, targetId);
+    
+    if (!draggedGroup || !targetGroup) return;
+
+    try {
+      if (isDescendant(draggedGroup, targetGroup)) {
+        toast({
+          title: "Invalid Move",
+          description: "Cannot move a group into its own descendant",
+          variant: "destructive",
+        });
+        return;
+      }
+
+      await moveGroup(draggedId, targetId);
+      toast({
+        title: "Success",
+        description: `Moved "${draggedGroup.name}" into "${targetGroup.name}"`,
+        variant: "success",
+      });
+      onRefresh();
+    } catch (error: any) {
+      toast({
+        title: "Error",
+        description: typeof error === 'string' ? error : (error?.message || "Failed to move group"),
+        variant: "destructive",
+      });
+    }
+  };
+
+  const openCreateDialog = (parentId: string | null) => {
+    setParentUuid(parentId);
+    setShowCreateDialog(true);
+  };
+
+  const openRenameDialog = (g: GroupData) => {
+    setRenameGroupUuid(g.uuid);
+    setRenameGroupName(g.name);
+    setRenameGroupIconId(g.icon_id ?? 48);
+    setShowRenameDialog(true);
+  };
+
+  const activeGroup = activeId ? findGroupByUuid(group, activeId) : null;
+
+  return (
+    <>
+      <DndContext
+        sensors={sensors}
+        collisionDetection={closestCenter}
+        onDragStart={handleDragStart}
+        onDragOver={handleDragOver}
+        onDragEnd={handleDragEnd}
+      >
+        <div className="h-full flex flex-col">
+          <div className="flex items-center justify-between border-b px-4 py-2">
+            <h2 className="text-sm font-semibold">Folders</h2>
+            <Button
+              variant="ghost"
+              size="icon"
+              className="h-7 w-7"
+              onClick={() => openCreateDialog(null)}
+            >
+              <Plus className="h-4 w-4" />
+            </Button>
+          </div>
+
+          <ContextMenu>
+            <ContextMenuTrigger asChild>
+              <ScrollArea className="flex-1">
+                <div className="p-1 min-h-full">
+                  <div
+                    className={`flex items-center gap-1 px-2 py-1.5 cursor-pointer rounded transition-colors ${
+                      selectedUuid === "_dashboard" ? "bg-accent" : "hover:bg-accent/50"
+                    }`}
+                    onClick={() => onSelectGroup("_dashboard")}
+                  >
+                    <LayoutDashboard className="h-4 w-4 text-muted-foreground flex-shrink-0" />
+                    <span className="truncate text-sm font-medium">Dashboard</span>
+                  </div>
+
+                  <div
+                    className={`flex items-center gap-1 px-2 py-1.5 cursor-pointer rounded transition-colors ${
+                      selectedUuid === "_favorites" ? "bg-accent" : "hover:bg-accent/50"
+                    }`}
+                    onClick={() => onSelectGroup("_favorites")}
+                  >
+                    <Star className="h-4 w-4 text-muted-foreground flex-shrink-0" />
+                    <span className="truncate text-sm font-medium">Favorites</span>
+                  </div>
+                  
+                  <DraggableFolder
+                    group={group}
+                    depth={0}
+                    selectedUuid={selectedUuid}
+                    expandedGroups={expandedGroups}
+                    overId={overId}
+                    onToggleExpand={toggleExpand}
+                    onSelectGroup={onSelectGroup}
+                    onOpenCreateDialog={openCreateDialog}
+                    onOpenRenameDialog={openRenameDialog}
+                    onDeleteGroup={handleDeleteGroup}
+                  />
+                </div>
+              </ScrollArea>
+            </ContextMenuTrigger>
+            <ContextMenuContent>
+              <ContextMenuItem onClick={() => openCreateDialog(null)}>
+                <Plus className="mr-2 h-4 w-4" />
+                New Group
+              </ContextMenuItem>
+            </ContextMenuContent>
+          </ContextMenu>
+        </div>
+        
+        <DragOverlay dropAnimation={null}>
+          {activeGroup ? (
+            <div className="flex items-center gap-1 px-2 py-1.5 bg-accent/50 rounded shadow-lg opacity-60">
+              {(() => {
+                const FolderIcon = getIconComponent(activeGroup.icon_id ?? 48);
+                return <FolderIcon className="h-4 w-4 text-muted-foreground" />;
+              })()}
+              <span className="text-sm font-medium">{activeGroup.name}</span>
+            </div>
+          ) : null}
+        </DragOverlay>
+      </DndContext>
+
+      <CreateGroupDialog
+        open={showCreateDialog}
+        onOpenChange={setShowCreateDialog}
+        groupName={newGroupName}
+        iconId={newGroupIconId}
+        onGroupNameChange={setNewGroupName}
+        onIconIdChange={setNewGroupIconId}
+        onSubmit={handleCreateGroup}
+      />
+
+      <RenameGroupDialog
+        open={showRenameDialog}
+        onOpenChange={setShowRenameDialog}
+        groupName={renameGroupName}
+        iconId={renameGroupIconId}
+        onGroupNameChange={setRenameGroupName}
+        onIconIdChange={setRenameGroupIconId}
+        onSubmit={handleRenameGroup}
+      />
+    </>
+  );
+}
