+++ components/main-app/hooks/useKeyboardShortcuts.ts
@@ -1,0 +1,21 @@
+"use client";
+
+import { useEffect } from "react";
+
+interface UseKeyboardShortcutsProps {
+  onSave: () => void;
+}
+
+export function useKeyboardShortcuts({ onSave }: UseKeyboardShortcutsProps) {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if ((e.ctrlKey || e.metaKey) && e.key === 's') {
+        e.preventDefault();
+        onSave();
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    return () => window.removeEventListener('keydown', handleKeyDown);
+  }, [onSave]);
+}
